implement DELETE method

+++ src/main/java/edu/oregonstate/languagesapi/controllers/CountryController.java
@@ -1,7 +1,9 @@
 package edu.oregonstate.languagesapi.controllers;
 
 import edu.oregonstate.languagesapi.models.Country;
+import edu.oregonstate.languagesapi.models.Language;
 import edu.oregonstate.languagesapi.services.CountryService;
+import edu.oregonstate.languagesapi.services.LanguageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +17,9 @@ public class CountryController {
 
     @Autowired
     CountryService countryService;
+
+    @Autowired
+    LanguageService languageService;
 
     @GetMapping("/countries")
     public ResponseEntity<List<Country>> get() {
@@ -36,6 +41,16 @@ public class CountryController {
 
     @DeleteMapping("/countries/{id}")
     public ResponseEntity<String> delete(@PathVariable("id") Long id) {
+        Country country = countryService.findById(id);
+        if (country == null) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+
+        List<Language> languages = languageService.findByCountry(id);
+        languages.forEach(language -> {
+            language.removeCountry(country);
+        });
+
         countryService.delete(id);
         return new ResponseEntity<String>("Country is deleted", HttpStatus.OK);
     }
